+++ .meta.php
@@ -18,18 +18,20 @@ namespace Illuminate\Database\Schema {
 
     use Illuminate\Support\Fluent;
     use Php\Support\Laravel\Database\Schema\Definitions\LikeDefinition;
-    use Php\Support\Laravel\Database\Schema\Definitions\UniqueDefinition;
-    use Php\Support\Laravel\Database\Schema\Definitions\PartialDefinition;
     use Php\Support\Laravel\Database\Schema\Definitions\ViewDefinition;
+    use Php\Support\Laravel\Database\Schema\Postgres\Builders\Indexes\PartialBuilder;
+    use Php\Support\Laravel\Database\Schema\Postgres\Builders\Indexes\Unique\UniqueBuilder;
 
     /**
      * @method LikeDefinition like(string $table)
      * @method Fluent ifNotExists()
-     * @method PartialDefinition partial($columns, ?string $index = null, ?string $algorithm = null)
-     * @method UniqueDefinition uniquePartial($columns, ?string $index = null, ?string $algorithm = null)
+     * @method PartialBuilder partial($columns, ?string $index = null, ?string $algorithm = null)
+     * @method UniqueBuilder uniquePartial($columns, ?string $index = null, ?string $algorithm = null)
      * @method ViewDefinition createView(string $view, string $select, bool $materialize = false)
      * @method ViewDefinition createViewOrReplace(string $view, string $select, bool $materialize = false)
-     * @method Fluent dropView(string $view)
+     * @method Fluent dropView(string $view, bool $materialize = false)
+     * @method Fluent dropViewIfExists(string $view, bool $materialize = false)
+     * @method Fluent ginIndex($columns, ?string $name = null)
      *
      * @mixin \Php\Support\Laravel\Database\Schema\Postgres\Blueprint
      */