+++ tests/Functional/Types/ArrayOfUuidTest.php
@@ -7,7 +7,7 @@ namespace Php\Support\Laravel\Database\Tests\Functional\Types;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use Php\Support\Laravel\Database\Schema\Postgres\Blueprint;
-use Php\Support\Laravel\Database\Schema\Postgres\Types\UuidArrayType;
+use Php\Support\Laravel\Database\Schema\Postgres\ColumnType;
 use Php\Support\Laravel\Database\Tests\AbstractTestCase;
 use Php\Support\Laravel\Database\Tests\Helpers\ColumnAssertions;
 use Php\Support\Laravel\Database\Tests\Helpers\IndexAssertions;
@@ -41,7 +41,7 @@ class ArrayOfUuidTest extends AbstractTestCase
             $definition->indexdef
         );
 
-        $this->assertTypeColumn('test_table', 'test_col', UuidArrayType::class);
+        $this->assertTypeColumn('test_table', 'test_col', ColumnType::UuidArray);
     }
 
 }