+++ tests/Functional/Types/IpNetworkTest.php
@@ -7,7 +7,7 @@ namespace Php\Support\Laravel\Database\Tests\Functional\Types;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use Php\Support\Laravel\Database\Schema\Postgres\Blueprint;
-use Php\Support\Laravel\Database\Schema\Postgres\Types\IpNetworkType;
+use Php\Support\Laravel\Database\Schema\Postgres\ColumnType;
 use Php\Support\Laravel\Database\Tests\AbstractTestCase;
 use Php\Support\Laravel\Database\Tests\Helpers\ColumnAssertions;
 use PHPUnit\Framework\Attributes\Test;
@@ -28,7 +28,7 @@ class IpNetworkTest extends AbstractTestCase
 
         static::assertTrue(Schema::hasTable('test_table'));
 
-        $this->assertTypeColumn('test_table', 'ip', IpNetworkType::class);
+        $this->assertTypeColumn('test_table', 'ip', ColumnType::IpNetwork);
 
         $value = '192.168/24';
         DB::insert(